💄 Added componets related to messages and message input
Four new componens where added.

The most relevant is <Message> that displays the message that a given user sent on the topic so it could be displayed on the <Chat> component.

The other components are related to each other, but <MessageInput> denotes the ability to send those messages on the topic.

In summary, <Message> displays the message and <MessageInput> sends the message